requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.6
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
huggingface_hub==0.19.4
python-environ==0.4.54
openai-whisper==20240930